refactor: move transaction modal into transactions workflow
